# openapi.yaml
---
openapi: 3.1.0
paths:
  /api:
    get:
      summary: Health test
      description: Health test endpoint
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                examples:
                - status: ok
      tags:
      - Status
  /api/v1/cbom/last/{limit}:
    get:
      summary: Return recently generated CBOMs from the repository
      description: Returns a list of the most recently generated CBOMs. The length
        of the list can by specified via the optional 'limit' parameter.
      parameters:
      - name: limit
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
      tags:
      - CBOM Resource
  /api/v1/cbom/scans:
    get:
      summary: Return a page of generated CBOMs from the repository
      description: "Returns one page of the most recently generated CBOMs, newest\
        \ first, wrapped in an object that also reports the total number of pages\
        \ so a pager can be rendered. The 1-based 'page' parameter defaults to 1 and\
        \ the 'limit' parameter defaults to 5 (capped at 100). Out-of-range values\
        \ are clamped rather than rejected."
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          format: int32
      - name: page
        in: query
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
      tags:
      - CBOM Resource
  /api/v1/cbom/{projectIdentifier}:
    post:
      parameters:
      - name: projectIdentifier
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
      summary: Store CBOM
      tags:
      - CBOM Resource
    get:
      parameters:
      - name: projectIdentifier
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
      summary: Get CBOM
      tags:
      - CBOM Resource
    delete:
      parameters:
      - name: projectIdentifier
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
      summary: Delete CBOM
      tags:
      - CBOM Resource
  /api/v1/compliance/check:
    get:
      summary: Verify the compliance of a stored CBOM identified by it's PURL against
        a policy
      description: "Returns the JSON sent by the Regulator API, containing various\
        \ information about the compliance of the CBOM for a set policy."
      parameters:
      - name: policyIdentifier
        in: query
        schema:
          type: string
      - name: projectIdentifier
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
      tags:
      - Compliance Resource
    post:
      summary: Verify the compliance of a provided CBOM against a policy
      description: "Returns the JSON sent by the Regulator API, containing various\
        \ information about the compliance of the CBOM for a set policy."
      parameters:
      - name: policyIdentifier
        in: query
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
      tags:
      - Compliance Resource
  /api/v1/scan:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Scan
      tags:
      - Scanning Resource
components:
  schemas:
    Credentials:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        pat:
          type: string
    ScanRequest:
      type: object
      properties:
        scanUrl:
          type: string
        branch:
          type: string
        subfolder:
          type: string
        credentials:
          $ref: "#/components/schemas/Credentials"
info:
  title: CBOMkit API
  version: 2.0.0-SNAPSHOT
  description: Service for generating and retrieving CBOMs
servers:
- url: http://localhost:8081
  description: Auto generated value
- url: http://0.0.0.0:8081
  description: Auto generated value